complete project

+++ src/main/java/com/illumina/shanqyeet/flashcarddemo/dtos/responses/PostCompleteGameResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PostCompleteGameResponse implements BaseResponse{
     private Integer totalScore;
-    private Double correctAnswerRate;
+    private Float correctAnswerRate;
     private Long averageAnswerTime;
 }
